+++ config/config.lua
@@ -1,0 +1,6 @@
+
+TICK_UPDATE_SQUAD_AI = 120
+DEFAULT_SQUAD_RADIUS = 10
+SOLDIER_MAX_AMMO = 100
+SQUAD_SIZE_MIN_BEFORE_HUNT = 5
+SQUAD_CHECK_RANGE = 25
