+++ data.lua
@@ -1,0 +1,7 @@
+require("prototypes.droiditem") -- any buildable or placable object/entity needs this 
+require("prototypes.droidbuildings")
+require("prototypes.droidentity") -- any buildable or placable object/entity needs this
+require("prototypes.droidrecipe")
+
+
+
