+++ info.json
@@ -1,0 +1,10 @@
+{
+
+	"name": "robotarmy",
+	"version": "0.1.3",
+	"title": "Robot Army",
+	"author": "Kyran F",
+	"contact": "kyranzor@gmail.com",
+	"homepage": "https://github.com/kyranf/robotarmyfactorio",
+	"description": "Manufacture and deploy automated robot soldiers to attack your foes!"
+}
