+++ locale/en/locale.cfg
@@ -1,0 +1,10 @@
+[item-name]
+droid = Battle Droid (undeployed)
+droid-dummy = Battle Droid (deployable)
+droid-assembling-machine = Droid Assembler
+[entity-name]
+droid = Battle Droid
+droid-assembling-machine = Droid Assembler
+[recipe-name]
+droid = Battle Droid (undeployed)
+droid-deploy = Battle Droid (Deployed)
