--- robotarmy_0.1.3/data.lua
@@ -1,7 +1,0 @@
-require("prototypes.droiditem") -- any buildable or placable object/entity needs this 
-require("prototypes.droidbuildings")
-require("prototypes.droidentity") -- any buildable or placable object/entity needs this
-require("prototypes.droidrecipe")
-
-
-
