+++ stdlib/core.lua
