+++ stdlib/entity/inventory.lua
